Fully working UltravioletSample12 app.
Had to work around a possible CoreCLR bug which was preventing the static constructor
for a control from running.

The "publishexe" target is still broken and needs to be fixed to actually copy Content
and other things intelligently into the Runtime directory rather than dumping everything
into the root of the folder.